+++ src/core/services/index.ts
@@ -1,5 +1,8 @@
 import ItemsService from './itemsService'
 import UsersService from './usersService'
 import RolesService from './rolesService'
+import Service from './service'
 
-export { ItemsService, UsersService, RolesService }
+export {
+  ItemsService, UsersService, RolesService, Service
+}