+++ src/core/services/itemsService.ts
@@ -1,5 +1,5 @@
 import { ItemsRepository } from '@repositories'
-import { Item } from '@src/core/models'
+import { Item } from '@models'
 import Service from './service'
 
 export default class ItemsService extends Service<Item, ItemsRepository> {